+++ backend/src/test/java/ca/bc/gov/backendstartapi/vo/parser/SmpMixVolumeTest.java
@@ -5,7 +5,6 @@ import static ca.bc.gov.backendstartapi.enums.parser.SmpMixHeader.POLLEN_VOLUME_
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrowsExactly;
 
-import java.math.BigDecimal;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 
@@ -15,18 +14,16 @@ class SmpMixVolumeTest extends CsvParsingResultTest {
   @Test
   void testFromMapMethod() {
     var vo = SmpMixVolume.fromMap(Map.of(PARENT_TREE_NUMBER, 1, POLLEN_VOLUME_ML, 2d));
-    assertEquals(new SmpMixVolume(1, BigDecimal.valueOf(2d)), vo);
+    assertEquals(new SmpMixVolume(1, 2d), vo);
   }
 
   @Test
   void parentTreeNumberMustBePositive() {
-    assertThrowsExactly(
-        IllegalArgumentException.class, () -> new SmpMixVolume(0, BigDecimal.valueOf(1d)));
+    assertThrowsExactly(IllegalArgumentException.class, () -> new SmpMixVolume(0, 1d));
   }
 
   @Test
   void pollenVolumeMustBeNonNegative() {
-    assertThrowsExactly(
-        IllegalArgumentException.class, () -> new SmpMixVolume(1, BigDecimal.valueOf(-1d)));
+    assertThrowsExactly(IllegalArgumentException.class, () -> new SmpMixVolume(1, -1d));
   }
 }
